typescript: fully implement typescript in ./api (backend) and ./client (frontend)
