+++ client/package.json
@@ -13,10 +13,10 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "dev": "react-scripts start",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "dev": "react-scripts start --typescript",
+    "start": "react-scripts start --typescript",
+    "build": "react-scripts build --typescript",
+    "test": "react-scripts test --typescript",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -36,5 +36,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
+    "typescript": "^5.7.3"
   }
 }
